move TrackerManager to own file & more efficient onNewStream implementation
